+++ src/grimp/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "3.16"
+__version__ = "3.17"
 
 from .application.graph import DetailedImport, Import, ImportGraph
 from .domain.analysis import PackageDependency, Route